fix(config): change LEVELS_EXPONENT from int to float
- Allow fractional values like 1.75 for levels exponent
- Update default from 2 to 2.0 for consistency
- Update config examples to reflect float type
- Fixes validation error when using fractional exponent values

+++ src/tux/shared/config/models.py
@@ -230,9 +230,9 @@ class XP(BaseModel):
         ),
     ]
     LEVELS_EXPONENT: Annotated[
-        int,
+        float,
         Field(
-            default=2,
+            default=2.0,
             description="Levels exponent",
             examples=[2, 3, 1.5],
         ),
